started implementing EFI support

+++ GentooInstallScript.sh
@@ -1,6 +1,6 @@
 EFI="false"
-install_drive="/dev/sda"
-swap_size="4G"
+install_drive="/dev/sdf"
+swap_size="128G"
 stage3_download="https://bouncer.gentoo.org/fetch/root/all/releases/x86/autobuilds/20210524T214502Z/stage3-i686-20210524T214502Z.tar.xz"
 timezone="Europe/Berlin"
 locales="en_US ISO-8859-1
@@ -9,13 +9,28 @@ de_DE ISO-8859-1
 de_DE.UTF-8 UTF-8"
 hostname="tux"
 
-
 if [ $1 = "--bootstrap" ]
 then
 
 	if [ $EFI = "true" ]
 	then
-		echo "test"
+		unalias cp
+		dd if=/dev/zero of=$install_drive bs=512 count=1
+		wipefs $install_drive
+		sfdisk $install_drive >> EOF
+label:gpt
+,256M
+,$swap_size 0x82
+;
+EOF
+		sfdisk -A "$install_drive"1
+		mkfs.vfat -F 32 "$install_drive"1
+		mkfs.ext4 "$install_drive"3
+		mkswap "$install_drive"2
+		swapon "$install_drive"2
+		mkdir /mnt/gentoo
+		mount "$install_drive"3 /mnt/gentoo
+		chmod 1777 /mnt/gentoo/tmp	
 	fi
 
 	if [ $EFI = "false" ]
@@ -23,18 +38,17 @@ then
 
 		#ensure that the GentooInstallScript is available in /mnt/gentoo
 		unalias cp
-		cp GentooInstallScript.sh /mnt/gentoo
 
 		#prepare file system
 		dd if=/dev/zero of=$install_drive bs=512 count=1
-		wipefs /dev/sda
+		wipefs $install_drive
 		sfdisk $install_drive << EOF
 label:dos
 ,256M
 ,$swap_size 0x82
 ;
 EOF
-		sfdisk -A $install_drive 1
+		sfdisk -A "$install_drive"1
 		mkfs.ext2 "$install_drive"1
 		mkfs.ext4 "$install_drive"3
 		mkswap "$install_drive"2
@@ -57,6 +71,7 @@ EOF
 	mount --rbind /dev /mnt/gentoo/dev
 	mount --make-rslave /mnt/gentoo/dev
 
+	cp GentooInstallScript.sh /mnt/gentoo
 	#chroot
 	chroot /mnt/gentoo /bin/bash
 
@@ -99,18 +114,24 @@ then
 	#installing sources and compiling kernel
 	emerge --ask sys-kernel/gentoo-sources
 	emerge --ask sys-kernel/genkernel
-	echo "/dev/sda1 /boot ext2 defaults 0 2" > /etc/fstab
+	echo ""$install_drive"1 /boot ext2 defaults 0 2" > /etc/fstab
 	genkernel all
 
 	#install firmware
 	emerge --ask sys-kernel/linux-firmware
 
 	#setup fstab
-	rm /etc/fstab
-	echo "/dev/sda1 /boot ext2 defaults 0 2\n" > /etc/fstab
-	echo "/dev/sda2 none swap sw 0 0\n" >> /etc/fstab
-	echo "/dev/sda3 / ext4 noatime 0 1\n" >> /etc/fstab
+	if [ $EFI=false ] 
+	then
+		rm /etc/fstab
+		echo ""$install_drive"1 /boot ext2 defaults 0 2\n" > /etc/fstab
+		echo ""$install_drive"2 none swap sw 0 0\n" >> /etc/fstab
+		echo ""$install_drive"3 ext4 noatime 0 1\n" >> /etc/fstab
+	fi
+	if [ $EFI=true ]
+	then
 
+	fi
 	#set hostname
 	echo "hostname=\"$hostname\"" > /etc/conf.d/hostname
 	
@@ -123,9 +144,7 @@ then
 	echo "127.0.0.1 tux" > /etc/hosts
 
 	#setup root password
-	echo "Enter root password:"
-	read rootpw
-	passwd $rootpw
+
 
 	#setup some tools
 	emerge --ask app-admin/sysklogd
@@ -140,9 +159,13 @@ then
 	then
 		#setup boot loader
 		emerge --ask --verbose sys-boot/grub:2
-		grub-install /dev/sda
+		grub-install $install_drive
 		grub-mkconfig -o /boot/grub/grub.cfg
 	fi
+	if [ $EFI = "true" ]
+		
+	fi
+	
 	
 
 fi
